refactor: `event` files to `common`
