+++ src/core/event/common/types.ts
